Makes sure transactions are properly committed or rolled back.
Depending on whether an exception occurred or not, a transaction will be committed or rolled back.  Previously, when the context manager has exited, `self.close_transaction()` was called, performing a commit on an only partially executed block.

This patch changes that to properly call `self.close_transaction(False)` when a failure happened, properly rolling back the entire transaction on the server end.